+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added GetDirectoryListing of a specific path
 - Added stderr to output of ShellExecute functions
 - Added ShellCmdExecute function
+- Added registry class with improved read/write functions
+- Added remote registry functions
 
 ### Changed
 - Updated mimikatz binaries
 - Changed mimikatz function to load in new thread, free input/output pointers
+- Updated registry tests
 
 ### Fixed
 - Fixed XML warning, removed angle brackets in comment